refactor(repository): move Revision/TreeReference to dialog-capability
Revision and TreeReference are plain serde data types (a logical clock
plus a search-tree root) with no dependency on the query engine, but
they lived in dialog-repository, which pulls dialog-query. A client that
only needs to name or deserialize a revision -- e.g. a wire DTO decoded
by a web page -- was forced to link the whole datalog engine.

Move both into dialog-capability, the light crate that already owns Did
and pulls no engine or storage/transport backend. TreeReference wraps a
bare [u8; 32] so its wire form is independent of any Blake3Hash wrapper.
dialog-repository re-exports both at their historical
dialog_repository::{Revision, TreeReference} paths, so every caller is
unchanged.
