+++ rust/dialog-capability/src/lib.rs
@@ -182,6 +182,9 @@ pub use selector::*;
 mod settings;
 pub use settings::*;
 
+mod revision;
+pub use revision::*;
+
 mod ability;
 pub use ability::*;
